refactor(stats): rename stats drawer components to *DrawerHost
MyStatsDrawer and StreakDrawer fetch their own data via use* hooks
and render the actual <Drawer>; rename to *Host to match the
convention introduced with MatchDrawerHost.
